requirement.txt
fastapi
uvicorn[standard]
firebase-admin
google-cloud-firestore
numpy